use gcc printf attribute
git-svn-id: https://svn.code.sf.net/p/sauerbraten/code/src@4729 07af1644-1164-4067-b897-63bf4dc9b748

+++ engine/sound.cpp
@@ -160,7 +160,7 @@ void initsound()
     if(Mix_OpenAudio(soundfreq, MIX_DEFAULT_FORMAT, 2, soundbufferlen)<0)
     {
         nosound = true;
-        conoutf(CON_ERROR, "sound init failed (SDL_mixer): %s", (size_t)Mix_GetError());
+        conoutf(CON_ERROR, "sound init failed (SDL_mixer): %s", Mix_GetError());
         return;
     }
 	Mix_AllocateChannels(soundchans);
